progress

+++ cgui/include/cassette/cgrid.h
@@ -25,7 +25,7 @@ typedef struct cgrid cgrid;
 
 [[nodiscard]] cgrid *cgrid_clone(const cgrid *gr);
 
-[[nodiscard]] cgrid *cgrid_create(uint32_t rows, uint32_t cols);
+[[nodiscard]] cgrid *cgrid_create(size_t rows, size_t cols);
 
 nullptr_t cgrid_destroy(cgrid *gr);
 
@@ -37,13 +37,13 @@ void cgrid_assign(cgrid *gr, cshell *sh);
 
 void cgrid_clear_warnings(cgrid *gr);
 
-void cgrid_flex_col(cgrid *gr, uint32_t col, double factor);
+void cgrid_flex_col(cgrid *gr, size_t col, double factor);
 
-void cgrid_flex_row(cgrid *gr, uint32_t row, double factor);
+void cgrid_flex_row(cgrid *gr, size_t row, double factor);
 
-void cgrid_resize_col(cgrid *gr, uint32_t col, int32_t size);
+void cgrid_resize_col(cgrid *gr, size_t col, int32_t size);
 
-void cgrid_resize_row(cgrid *gr, uint32_t row, int32_t size);
+void cgrid_resize_row(cgrid *gr, size_t row, int32_t size);
 
 void cgrid_retire(cgrid *gr);
